# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Authentication Manager API
  description: |
    Production-ready backend API for Unified Authentication Manager.
    
    Features:
    - Zero-knowledge vault (client-side encryption)
    - Device-based MFA with cryptographic signatures
    - Email relay with encrypted routing
    - JWT authentication with refresh tokens
    - Rate limiting and security headers
  version: 1.0.0
  contact:
    name: UAM API Support
    url: https://api.uam.example
servers:
  - url: https://api.uam.example/api/v1
    description: Production server
  - url: http://localhost:5000/api/v1
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
    RegisterRequest:
      type: object
      required: [email, password, masterKeySalt, masterKeyParams, kdfIterations]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        masterKeySalt:
          type: string
          minLength: 16
        masterKeyParams:
          type: object
        kdfIterations:
          type: integer
          minimum: 100000
          maximum: 600000
          default: 200000
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        deviceId:
          type: string
          format: uuid
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    DeviceRegisterRequest:
      type: object
      required: [name, publicKey, platform]
      properties:
        name:
          type: string
          minLength: 2
        publicKey:
          type: string
          description: Base64-encoded Ed25519 or secp256k1 public key
        keyAlgorithm:
          type: string
          enum: [ed25519, secp256k1]
          default: ed25519
        fcmTokenCipher:
          type: string
          description: Encrypted FCM token
        platform:
          type: string
          enum: [android, ios, web, desktop]
        appVersion:
          type: string
...
    AliasCreateRequest:
      type: object
      properties:
        routingMode:
          type: string
          enum: [DIRECT, HMAC_LOOKUP]
          default: DIRECT
        targetCiphertext:
          type: string
          description: Encrypted destination email (required for DIRECT)
        targetHmac:
          type: string
          description: HMAC for lookup (required for HMAC_LOOKUP)
        hmacKeyVersion:
          type: integer
          minimum: 1
paths:
...
  /auth/register:
    post:
      tags: [auth]
      summary: Register new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  sessionId:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      email:
                        type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/logout:
    post:
      tags: [auth]
      summary: Logout and revoke session
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices:
    get:
      tags: [devices]
      summary: List user devices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of devices
          content:
            application/json:
              schema:
                type: object
                properties:
                  devices:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [devices]
      summary: Register new device
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceRegisterRequest'
      responses:
        '201':
          description: Device registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  device:
                    type: object
        '401':
          $ref: '#/components/responses/Unauthorized'
  /devices/{id}:
    delete:
      tags: [devices]
      summary: Revoke device
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Device revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /aliases:
    get:
      tags: [aliases]
      summary: List email aliases
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of aliases
          content:
            application/json:
              schema:
                type: object
                properties:
                  aliases:
                    type: array
                    items:
                      type: object
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [aliases]
      summary: Create email alias
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AliasCreateRequest'
      responses:
        '201':
          description: Alias created
          content:
            application/json:
              schema:
                type: object
                properties:
                  alias:
                    type: object
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /aliases/{id}:
    delete:
      tags: [aliases]
      summary: Revoke email alias
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Alias revoked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...